Removing batch estimation. Tqdm less precise but lighter code
